requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.3
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
cryptography==42.0.8
